Handle non-set TimeInfo Read/Write fixing Clock In bug
The bug was caused by the unhandled case of an uninitialized TimeInfo written
to a file and read back. The read back version was initialized to 0 sec in year
1970. The LaborDB::IsUserOnClock() checks the end TimeInfo to be uninizialized.
With this bug the employee is never clocked in after a restart of viewtouch.

Fix this by writing out non-set TimeInfo values as sec==0 year==0, and when
handle them accordingly in the Read(TimeInfo) function.

Add tests for TimeInfo and a regression test for this bug to never again
creep in unnoticed.

Fixes: https://github.com/ViewTouch/viewtouch/issues/108