main: report: Handle nullptr const char ptr in Report::Text again
In 9f10c00e461f6b86c8f9c2f9c058698df228efc7 main: report: remove custom DList, updates
the `Report::Text` function was switched to accept `std::string&` instead
of `const char*`. In the conversion the possibility of the char ptr to
be a nullptr was forgotten. A nullptr for text is interpreted as
`draw_a_line`. As `std::string` has no distinction between nullptr and
an empty string introduce the variable `bool draw_a_line` and
reintroduce the `const char*` overloads for `Report::Text` and
`ReportEntry()`.

Add tests to check ReportEntry constructors with `const char *` and
`std::string`. Especially the new behavior when `nullptr` is passed as
text, which should be the only instance `draw_a_line` should be set to
true.

fixes: https://github.com/ViewTouch/viewtouch/issues/115