time_info: fix Set(string) parsing add unittest, C++20 compatibility
The format string passed to `date::parse()` was wrong, and thus never
succeeded to parse a date string. Fix it such that the date is actually
parsed.

And then actually use the parsed date info for TimeInfo. Previously it was
parsed, but never stored in the class.

This also fixes gcc-14 C++20 compatibility.

Affected code is in `manager.cc` the function `RunReport()`

```cpp
        // need to pull out "Report From To"
        // date will be in the format "DD/MM/YY,HH:MM" in 24hour format
        if (NextToken(report_name, report_string, ' ', &idx))
        {
            if (NextToken(report_from, report_string, ' ', &idx))
            {
                from.Set(report_from);
                if (NextToken(report_to, report_string, ' ', &idx))
                    to.Set(report_to);
            }
        }
        if (!from.IsSet())
        {  // set date to yesterday morning, 00:00
            from.Set();
            from -= date::days(1);
            from.Floor<date::days>();
        }
        if (!to.IsSet())
        {  // set date to last night, 23:59
            to.Set();
            to.Floor<date::days>();
            to -= std::chrono::seconds(1);
        }
```

So the date was always from yesterday morning to last night.

Add unittest to check this function in the future.

Fixes: https://github.com/ViewTouch/viewtouch/issues/145